Move test_st_text to new file
